% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_load_expression.R
\name{load_expression}
\alias{load_expression}
\title{Load Expression Data from TSS BED File}
\usage{
load_expression(file, genome = NULL)
}
\arguments{
\item{file}{Path to a BED-format file with at least 3 columns: chromosome, TSS, and expression.}

\item{genome}{Optional \link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo} object to restrict chromosomes and assign seqinfo.}
}
\value{
A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object with 1-bp ranges and a metadata column \code{expression}.
}
\description{
Load a BED-format file containing transcription start site (TSS) coordinates and expression values,
and convert it into a \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object with an "expression" metadata column.
}
\examples{
expr_file <- system.file("extdata", "expression_tss.bed", package = "chromcall")
genome_file <- system.file("extdata", "genome.txt", package = "chromcall")
genome <- load_genome(genome_file)
expr_gr <- load_expression(expr_file, genome = genome)
expr_gr

}
