% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_load_windows.R
\name{load_windows}
\alias{load_windows}
\title{Load Promoter Window Regions from a BED File}
\usage{
load_windows(file, genome, blacklist = NULL)
}
\arguments{
\item{file}{Path to a BED-format file specifying genomic windows (e.g., promoters).}

\item{genome}{A \link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo} object used to define the genome context.}

\item{blacklist}{Optional \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object. If provided, each region
will be annotated with a logical flag \code{blacklist = TRUE} if it overlaps any blacklist region.}
}
\value{
A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object. If \code{blacklist} is provided, it includes an
additional metadata column \code{blacklist}.
}
\description{
Load genomic regions (e.g., promoter windows) from a BED-format file and associate
them with a given \link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo} genome object. If a blacklist is provided,
an additional logical column \code{blacklist} is added to indicate whether each region
overlaps with any blacklisted region.
}
\examples{
genome_file <- system.file("extdata", "genome.txt", package = "chromcall")
genome <- load_genome(genome_file)
blacklist_file <- system.file("extdata", "blacklist.bed", package = "chromcall")
blacklist <- load_bedfile(blacklist_file, genome = genome)
promoter_file <- system.file("extdata", "example.bed", package = "chromcall")
promoters <- load_windows(promoter_file, genome, blacklist = blacklist)
promoters

}
